+++ api/rest/handler_test.go
@@ -1,4 +1,4 @@
-package handler
+package rest
 
 import (
 	"testing"
